deadcode: groupbyRSet
